+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 - Use the NFO as description if the description is empty
+- Message API
+    - `GET /api/v1/message/messages` get all messages in a folder.
+    - `GET /api/v1/message/unread` get all unread messages in a folder.
+    - `GET /api/v1/message/read` get / read a message.
+    - `POST /api/v1/message/mark_read` mark messages as read.
+    - `POST /api/v1/message/delete` delete messages.
+    - `POST /api/v1/message/send` send a new message.
 
 ### Changed
 - `Template::render()` now returns `HttpResponse` instead of `Template`